Bring auth logic to lib folder, fix redirect errors
