+++ front-end/src/app/auth/[...nextauth]/route.ts
@@ -1,2 +1,2 @@
-import { handlers } from "@/auth";
+import { handlers } from "@/lib/auth";
 export const { GET, POST } = handlers;
