+++ front-end/src/app/page.tsx
@@ -1,5 +1,5 @@
-import { auth } from "@/auth";
 import { LoginForm } from "@/components/Auth/LoginForm";
+import { auth } from "@/lib/auth";
 
 export default async function Home() {
   const session = await auth();
