+++ front-end/src/middleware.ts
@@ -1,1 +1,1 @@
-export { auth as middleware } from "@/auth";
+export { auth as middleware } from "@/lib/auth";
